AvailableStorage: return int64, add tests

+++ pkg/aws/instance_info.go
@@ -22,11 +22,11 @@ type Instance struct {
 	InstanceStorageDeviceSize int64
 }
 
-func (i Instance) AvailableStorage(instanceStorageScaleFactor float64, rootVolumeSize int64, rootVolumeScaleFactor float64) float64 {
+func (i Instance) AvailableStorage(instanceStorageScaleFactor float64, rootVolumeSize int64, rootVolumeScaleFactor float64) int64 {
 	if i.InstanceStorageDevices == 0 {
-		return float64(rootVolumeSize) * rootVolumeScaleFactor
+		return int64(float64(rootVolumeSize) * rootVolumeScaleFactor)
 	}
-	return instanceStorageScaleFactor * float64(i.InstanceStorageDevices*i.InstanceStorageDeviceSize)
+	return int64(instanceStorageScaleFactor * float64(i.InstanceStorageDevices*i.InstanceStorageDeviceSize))
 }
 
 type instanceInfo struct {
